v0.1.12 — reorganized settings, always-visible filter pickers
Settings → Filters is regrouped into clear SCOPE / HUNTS / DISPLAY
sections. The world and expansion pickers are now always-visible
framed panels with All / None buttons and a live status line — no
more hiding the list behind an enable checkbox.

The two near-identical picker blocks are unified into one shared
FilterPanel helper; the repeated section-header pattern across every
tab is unified into a Section() helper. WorldWhitelist is now
List<int> (identical JSON on disk — existing configs load unchanged)
so one helper drives both pickers. Default state is still "no filter"
so existing installs are unaffected; an enabled flag disambiguates
the empty whitelist (None = mute all, explicit).

+++ FaloopMessenger/FaloopSocketClient.cs
@@ -474,7 +474,7 @@ public class FaloopSocketClient : IDisposable
 
         // Per-world filter (a subset of the DC the user explicitly ticked).
         // Only applies when enabled; the empty/disabled case keeps the full DC.
-        if (_config.WorldFilterEnabled && !_config.WorldWhitelist.Contains(worldId))
+        if (_config.WorldFilterEnabled && !_config.WorldWhitelist.Contains((int)worldId))
             return;
 
         // The nested "data" object holds spawn-specific fields (Spawn record)
